Move SchemaType to config_keywords.py & use Enum.auto()
